requirements.txt
redis==5.0.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
ruamel.yaml==0.18.5
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1